changed all file pathing to have seperated headers and implementations
