+++ drive/CMakeLists.txt
@@ -20,7 +20,8 @@ find_package(libhal-esp8266 REQUIRED CONFIG)
 add_executable(${PROJECT_NAME}
     main.cpp
     applications/application.cpp
-    platforms/${platform}.cpp)
+    platforms/${platform}.cpp
+    src/drc_speed_sensor.hpp)
 target_include_directories(${PROJECT_NAME} PUBLIC .)
 target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_20)
 target_link_options(${PROJECT_NAME} PRIVATE
